Render tree impostors and split presentation modules
