+++ crates/adventuresim-tactical-client/src/presentation/environment.rs
@@ -1,0 +1,159 @@
+use super::*;
+
+pub(super) fn scene_sunlight_illuminance(environment: &SceneEnvironment) -> f32 {
+    let intensity = f32::from(environment.weather.intensity_bps) / 10_000.0;
+    let transmission = match environment.weather.precipitation {
+        Precipitation::Clear => 1.0,
+        Precipitation::Rain => 0.62 - intensity * 0.27,
+        Precipitation::Snow => 0.72 - intensity * 0.22,
+    };
+    lux::DIRECT_SUNLIGHT * transmission.clamp(0.25, 1.0)
+}
+
+pub(super) fn scene_distance_fog(environment: &SceneEnvironment) -> DistanceFog {
+    let intensity = f32::from(environment.weather.intensity_bps) / 10_000.0;
+    let (start, end, color) = match environment.weather.precipitation {
+        Precipitation::Clear => (30_000.0, 50_000.0, Color::srgb_u8(188, 201, 207)),
+        Precipitation::Rain => (
+            800.0 + 3_500.0 * (1.0 - intensity),
+            3_000.0 + 14_000.0 * (1.0 - intensity),
+            Color::srgb_u8(112, 126, 135),
+        ),
+        Precipitation::Snow => (
+            600.0 + 2_500.0 * (1.0 - intensity),
+            2_500.0 + 10_000.0 * (1.0 - intensity),
+            Color::srgb_u8(207, 216, 220),
+        ),
+    };
+    DistanceFog {
+        color,
+        falloff: FogFalloff::Linear { start, end },
+        ..default()
+    }
+}
+
+#[derive(Component)]
+pub(crate) struct TacticalSunlight;
+
+pub(in crate::presentation) fn setup_tactical_presentation(
+    mut commands: Commands,
+    mut scattering_mediums: ResMut<Assets<ScatteringMedium>>,
+    settings: Res<TacticalGraphicsSettings>,
+) {
+    commands.spawn((
+        Name::new("Tactical sunlight"),
+        TacticalSunlight,
+        Transform::from_xyz(200.0, 1000.0, 100.0).looking_at(Vec3::ZERO, Vec3::Y),
+        DirectionalLight {
+            shadow_maps_enabled: settings.shadows_enabled,
+            illuminance: lux::DIRECT_SUNLIGHT,
+            ..default()
+        },
+    ));
+
+    if settings.atmosphere_enabled {
+        commands.spawn(Atmosphere::earth(
+            scattering_mediums.add(ScatteringMedium::default()),
+        ));
+    }
+
+    let mut camera = commands.spawn((
+        Name::new("Tactical gameplay camera"),
+        Camera3d::default(),
+        Projection::Perspective(PerspectiveProjection {
+            fov: 80.0_f32.to_radians(),
+            far: 60_000.0,
+            ..default()
+        }),
+        Exposure::SUNLIGHT,
+        Tonemapping::AcesFitted,
+        DistanceFog {
+            color: Color::srgb_u8(188, 201, 207),
+            falloff: FogFalloff::Linear {
+                start: 30_000.0,
+                end: 50_000.0,
+            },
+            ..default()
+        },
+        // Gameplay MSAA is deliberately off even in the full preset.
+        Msaa::Off,
+    ));
+    if settings.atmosphere_enabled {
+        camera.insert(AtmosphereSettings::default());
+        if settings.environment_light_enabled {
+            camera.insert(AtmosphereEnvironmentMapLight {
+                size: UVec2::splat(settings.environment_map_size),
+                ..default()
+            });
+        }
+    }
+    if settings.bloom_enabled {
+        camera.insert(Bloom::NATURAL);
+    }
+    if settings.ssao_enabled {
+        camera.insert(ScreenSpaceAmbientOcclusion::default());
+    }
+}
+
+pub(super) fn on_environment_added(
+    event: On<Add, SceneEnvironment>,
+    environments: Query<&SceneEnvironment>,
+    mut sunlight: Single<&mut DirectionalLight, With<TacticalSunlight>>,
+    mut fog: Single<&mut DistanceFog, With<Camera3d>>,
+) -> Result {
+    let environment = environments.get(event.entity)?;
+    sunlight.illuminance = scene_sunlight_illuminance(environment);
+    **fog = scene_distance_fog(environment);
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn environment(precipitation: Precipitation, intensity_bps: u16) -> SceneEnvironment {
+        SceneEnvironment {
+            scene_digest: "fixture".into(),
+            generation_version: TACTICAL_SCENE_GENERATION_VERSION,
+            weather: WeatherSnapshot {
+                rules_version: WEATHER_RULES_VERSION,
+                interval_start_minute: 0,
+                cell_latitude: 0,
+                cell_longitude: 0,
+                temperature_deci_c: 100,
+                wind_speed_bps: 0,
+                precipitation,
+                intensity_bps,
+                ground_moisture_bps: 0,
+                snow_cover_bps: 0,
+            },
+            canopy_bps: 0,
+            wetland_bps: 0,
+            cultivation_bps: 0,
+            water_bps: 0,
+            hilly_bps: 0,
+        }
+    }
+
+    #[test]
+    fn precipitation_dims_but_never_extinguishes_sunlight() {
+        let clear = scene_sunlight_illuminance(&environment(Precipitation::Clear, 0));
+        let rain = scene_sunlight_illuminance(&environment(Precipitation::Rain, 10_000));
+        let snow = scene_sunlight_illuminance(&environment(Precipitation::Snow, 10_000));
+        assert!(rain < snow && snow < clear);
+        assert!(rain >= lux::DIRECT_SUNLIGHT * 0.25);
+    }
+
+    #[test]
+    fn heavy_weather_reduces_visibility_below_clear_horizon() {
+        let clear = scene_distance_fog(&environment(Precipitation::Clear, 0));
+        let rain = scene_distance_fog(&environment(Precipitation::Rain, 10_000));
+        let snow = scene_distance_fog(&environment(Precipitation::Snow, 10_000));
+        let fog_end = |fog: &DistanceFog| match &fog.falloff {
+            FogFalloff::Linear { end, .. } => *end,
+            _ => panic!("scene weather must use bounded linear fog"),
+        };
+        assert!(fog_end(&rain) < fog_end(&clear));
+        assert!(fog_end(&snow) < fog_end(&clear));
+    }
+}
